+++ Packs/Base/Scripts/CommonServerPython/CHANGELOG.md
@@ -2,6 +2,10 @@
   - Added retry mechanism to the BaseClient.
   - Fixed an issue where the **appendContext** function did not behave as expected.
 
+
+## [20.5.0] - 2020-05-12
+  - Added retry mechanism to the BaseClient
+
 ## [20.4.1] - 2020-04-29
   - Deprecated the following enums: 
     - **entryTypes**